% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TetrominoAction.R
\name{checkNext3Block_y}
\alias{checkNext3Block_y}
\title{Check tetrominos moving towards}
\usage{
checkNext3Block_y(cubes, tables, direct)
}
\arguments{
\item{cubes}{a matrix which contains the information where tetromino locates}

\item{tables}{a matrix which contains the information of the current backgroup including previous tetrominos location}
}
\description{
Check if the tetromino should move down 3
}
\details{
This function will check if the tetromino can move down 3
}
\author{
Tingting & Chang
}
