% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TetrominoAction.R
\name{checkNextBlock_x}
\alias{checkNextBlock_x}
\title{Check tetrominos moving right/left}
\usage{
checkNextBlock_x(cubes, tables, direct)
}
\arguments{
\item{cubes}{a matrix which contains the information where tetromino locates}

\item{tables}{a matrix which contains the information of the current backgroup including previous tetrominos location}
}
\description{
Check if the tetromino should move left or right
}
\details{
This function will check if the tetromino can move right ot left.
}
\author{
Tingting & Chang
}
