% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TetrominoAction.R
\name{checkNextBlock_y}
\alias{checkNextBlock_y}
\title{Check tetrominos moving towards}
\usage{
checkNextBlock_y(cubes, tables)
}
\arguments{
\item{cubes}{a matrix which contains the information where tetromino locates}

\item{tables}{a matrix which contains the information of the current backgroup including previous tetrominos location}
}
\description{
Check if the tetromino should move down
}
\details{
This function will check if the tetromino can move down
}
\author{
Tingting & Chang
}
