% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableAction.R
\name{drawTable}
\alias{drawTable}
\title{Backgroup initialization}
\usage{
drawTable()
}
\description{
Backgroup initialization
}
\details{
This function will draw a blank table in Website before starting game
}
\author{
Tingting & Chang
}
