% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameAction.R
\name{endGame}
\alias{endGame}
\title{End Game}
\usage{
endGame(tables)
}
\arguments{
\item{tables}{a matrix which contains the information where tetrominos locates}
}
\description{
Check if the game will ends
}
\details{
This function will check if any coloums is filled by same color and the game will be end
}
\author{
Tingting
}
