% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableID.R
\name{GetIndexList}
\alias{GetIndexList}
\title{Get index of colored square in a tetromino}
\usage{
GetIndexList(tetromino)
}
\arguments{
\item{tetromino}{the matrix indicate the location of a tetromino}
}
\description{
Get (x,y) for squares in a tetromino which should be colored.
}
\details{
This function will write a 3x3 matrix as a specific shape tetromino.
}
\author{
Tingting & Chang
}
