% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableID.R
\name{Getinittetromino}
\alias{Getinittetromino}
\title{Generate tetromino}
\usage{
Getinittetromino()
}
\description{
Generate a random tetromino without any shape
}
\details{
This function will generate a random tetromino.
}
\author{
Tingting & Chang
}
