% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameAction.R
\name{GetScore}
\alias{GetScore}
\title{Win points}
\usage{
GetScore(tables)
}
\arguments{
\item{tables}{a matrix which contains the information where tetrominos locates}
}
\description{
Check if the player will win points
}
\details{
This function will check if any lines are filled by same color and player will win 10 points
}
\author{
Tingting & Chang
}
