% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableID.R
\name{GetType}
\alias{GetType}
\title{Shape tetromino}
\usage{
GetType(tetromino)
}
\arguments{
\item{tetromino}{the matrix indicate the location of a tetromino}
}
\description{
assign a shape to tetromino
}
\details{
This function will assign a shape to tetromino.
}
\author{
Tingting & Chang
}
