% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableID.R
\name{GetTypeMatrix}
\alias{GetTypeMatrix}
\title{Write tetromino information}
\usage{
GetTypeMatrix(type)
}
\arguments{
\item{type}{a indicate for whichi shape will be chosen.}
}
\description{
Write a 3x3 matrix as a specific shape tetromino
}
\details{
This function will write a 3x3 matrix as a specific shape tetromino.
}
\author{
Tingting & Chang
}
