% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableID.R
\name{GetxyforBoxes}
\alias{GetxyforBoxes}
\title{Get location}
\usage{
GetxyforBoxes(BoxIDx, BoxIDy)
}
\arguments{
\item{BoxIDx}{the ID for X-axis}

\item{BoxIDy}{the ID for y-axis}
}
\description{
Get accuracy location (x, y) for each square for given xid and yid
}
\details{
This function will get accuracy lcoation (x, y) for each square for given xid and yid.
}
\author{
Tingting & Chang
}
