% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TetrominoAction.R
\name{MoveDown}
\alias{MoveDown}
\title{Move right}
\usage{
MoveDown(cubes, tables)
}
\arguments{
\item{cubes}{a matrix which contains the information where tetromino locates}

\item{tables}{a matrix which contains the information of the current backgroup including previous tetrominos location}
}
\description{
Move down the dropping tetromino.
}
\details{
This function will move  down the dropping tetromino.
}
\author{
Tingting & Chang
}
