% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TetrominoAction.R
\name{MoveLeft}
\alias{MoveLeft}
\title{Move left}
\usage{
MoveLeft(cubes, tables)
}
\arguments{
\item{cubes}{a matrix which contains the information where tetromino locates}

\item{tables}{a matrix which contains the information of the current backgroup including previous tetrominos location}
}
\description{
Move the dropping tetromino to the left
}
\details{
This function will move the dropping tetromino to the left
}
\author{
Tingting & Chang
}
