% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TetrominoAction.R
\name{rotate}
\alias{rotate}
\title{Rotate the dropping tetromino.}
\usage{
rotate(cubes, tables)
}
\arguments{
\item{cubes}{a matrix which contains the information where tetromino locates}
}
\description{
Rotate the dropping tetromino.
}
\details{
This function will rotate the dropping tetromino.
}
\author{
Tingting & Chang
}
