% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableID.R
\name{totalMatrix}
\alias{totalMatrix}
\title{Generate backgroud matrix}
\usage{
totalMatrix()
}
\description{
Generate a NA 20x10 matrix for the game
}
\details{
This function will generate a NA 20x10 matrix for the game.
}
\author{
Tingting & Chang
}
