% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableAction.R
\name{updateBackGround}
\alias{updateBackGround}
\title{Update location of tetriminos}
\usage{
updateBackGround(TableMatrix)
}
\arguments{
\item{TableMatrix}{the matrix indicates will part should be colored as blue}
}
\description{
Update the whole sreen of Teris
}
\details{
This function will update the figure when location of tetriminos is change.
when the location of tetriminos is changed, the variable TableMatrix will change as well.
}
\author{
Tingting & Chang
}
