% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableAction.R
\name{UpdateTable}
\alias{UpdateTable}
\title{Update location of tetriminos}
\usage{
UpdateTable(gp, tetromino)
}
\arguments{
\item{gp}{plot object which contian current backgroup}

\item{tetromino}{the matrix indicates the location of dropping tetromino}
}
\description{
Update Teris figure when a tetromino is dropping
}
\details{
This function will update the figure when location of tetriminos is dropping or rotating.
}
\author{
Tingting & Chang
}
